third_party/WebKit/LayoutTests/http/tests/inspector/resources/clear-storage-view-expected.txt

CONSOLE MESSAGE: line 112: InspectorTest.IndexedDB_callback1
CONSOLE MESSAGE: line 112: InspectorTest.IndexedDB_callback2
CONSOLE MESSAGE: line 112: InspectorTest.IndexedDB_callback3
Tests Clear Storage view functions.

Tree element found: true
Clear storage view is visible: true
0 B used out of - temporary storage quota
0 B used out of - persistent storage quota

Running: Now with data
9.5 MB used out of - temporary storage quota
0 B used out of - persistent storage quota
